# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Poseidon API Specification
  description: |-
    A simple API for our financial aggregator app Poseidon.
    Project link:
    - [Poseidon Github Page](https://github.com/Dave-log/Poseidon)
  version: '1.0.0'
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /bidList/validate:
    post:
      summary: Validate and save a new bid list
      tags:
        - Bid List
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidList'
      responses:
        '302':
          description: Bid list saved successfully and redirected to list page
        '400':
          description: Invalid request, bid list not saved
        '500':
          description: Internal server error
  /bidList/update/{id}:
    get:
      summary: Display form for updating a bid list
      tags:
        - Bid List
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the bid list to update
          schema:
            type: integer
      responses:
        '200':
          description: Update form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Update bid list form</body></html>"

    post:
      summary: Update a bid list
      tags:
        - Bid List
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the bid list to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidList'
      responses:
        '302':
          description: Bid list updated successfully and redirected to list page
        '400':
          description: Invalid request, bid list not updated
        '500':
          description: Internal server error
  /bidList/delete/{id}:
    get:
      summary: Delete a bid list
      tags:
        - Bid List
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the bid list to delete
          schema:
            type: integer
      responses:
        '302':
          description: Bid list deleted successfully and redirected to list page
        '404':
          description: Bid list not found
        '500':
          description: Internal server error
...
  /curvePoint/add:
    post:
      summary: Add a new curve point
      tags:
        - Curve Point
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurvePoint'
      responses:
        '302':
          description: Curve point added successfully and redirected to list page
        '400':
          description: Invalid request, curve point not added
        '500':
          description: Internal server error
  /curvePoint/update/{id}:
    get:
      summary: Display form for updating a curve point
      tags:
        - Curve Point
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the curve point to update
          schema:
            type: integer
      responses:
        '200':
          description: Update form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Update curve point form</body></html>"
    post:
      summary: Update a curve point
      tags:
        - Curve Point
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the curve point to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurvePoint'
      responses:
        '302':
          description: Curve point updated successfully and redirected to list page
        '400':
          description: Invalid request, curve point not updated
        '500':
          description: Internal server error
  /curvePoint/delete/{id}:
    get:
      summary: Delete a curve point
      tags:
        - Curve Point
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the curve point to delete
          schema:
            type: integer
      responses:
        '302':
          description: Curve point deleted successfully and redirected to list page
        '404':
          description: Curve point not found
        '500':
          description: Internal server error
...
  /rating/add:
    get:
      summary: Display form for adding a new rating
      tags:
        - Rating
      responses:
        '200':
          description: Add rating form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Add rating form</body></html>"
    post:
      summary: Add a new rating
      tags:
        - Rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
      responses:
        '302':
          description: Rating added successfully and redirected to list page
        '400':
          description: Invalid request, rating not added
        '500':
          description: Internal server error
  /rating/update/{id}:
    get:
      summary: Display form for updating a rating
      tags:
        - Rating
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the rating to update
          schema:
            type: integer
      responses:
        '200':
          description: Update form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Update rating form</body></html>"
    post:
      summary: Update a rating
      tags:
        - Rating
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the rating to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
      responses:
        '302':
          description: Rating updated successfully and redirected to list page
        '400':
          description: Invalid request, rating not updated
        '500':
          description: Internal server error
  /rating/delete/{id}:
    get:
      summary: Delete a rating
      tags:
        - Rating
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the rating to delete
          schema:
            type: integer
      responses:
        '302':
          description: Rating deleted successfully and redirected to list page
        '404':
          description: Rating not found
        '500':
          description: Internal server error
...
  /ruleName/add:
    get:
      summary: Display form for adding a new rule name
      tags:
        - Rule Name
      responses:
        '200':
          description: Add rule name form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Add rule name form</body></html>"
    post:
      summary: Add a new rule name
      tags:
        - Rule Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleName'
      responses:
        '302':
          description: Rule name added successfully and redirected to list page
        '400':
          description: Invalid request, rule name not added
        '500':
          description: Internal server error
  /ruleName/update/{id}:
    get:
      summary: Display form for updating a rule name
      tags:
        - Rule Name
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the rule name to update
          schema:
            type: integer
      responses:
        '200':
          description: Update form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Update rule name form</body></html>"
    post:
      summary: Update a rule name
      tags:
        - Rule Name
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the rule name to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleName'
      responses:
        '302':
          description: Rule name updated successfully and redirected to list page
        '400':
          description: Invalid request, rule name not updated
        '500':
          description: Internal server error
  /ruleName/delete/{id}:
    get:
      summary: Delete a rule name
      tags:
        - Rule Name
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the rule name to delete
          schema:
            type: integer
      responses:
        '302':
          description: Rule name deleted successfully and redirected to list page
        '404':
          description: Rule name not found
        '500':
          description: Internal server error
...
components:
  schemas:
    BidList:
      type: object
      properties:
        account:
          type: string
        type:
          type: string
        bidQuantity:
          type: number
      required:
        - account
        - type
        - bidQuantity
    CurvePoint:
      type: object
      properties:
        curveId:
          type: integer
        term:
          type: number
        value:
          type: number
      required:
        - curveId
        - term
        - value
...
    Rating:
      type: object
      properties:
        moodysRating:
          type: string
        sandPRating:
          type: string
        fitchRating:
          type: string
        orderNumber:
          type: integer
    RuleName:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        json:
          type: string
        template:
          type: string
        sqlStr:
          type: string
        sqlPart:
          type: string
      required:
        - name
        - json
        - template
        - sqlPart
...